# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  DentistPatch:
    properties:
      last_name:
        type: string
      license:
        type: string
      name:
        type: string
    type: object
  DentistPost:
    properties:
      last_name:
        type: string
      license:
        type: string
      name:
        type: string
    required:
    - last_name
    - license
    - name
    type: object
  DentistPut:
    properties:
      last_name:
        type: string
      license:
        type: string
      name:
        type: string
    required:
    - last_name
    - license
    - name
    type: object
  DentistResponse:
    properties:
      id:
        type: integer
      last_name:
        type: string
      license:
        type: string
      name:
        type: string
    type: object
  ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      path:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
...
info:
  contact:
    email: gabriela.calicanton@gmail.com
    name: Gabriela Cecilia Calicanton
    url: http://www.swagger.io/support
  description: API of Dental Clinic
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Dental Clinic API
  version: "2.0"
paths:
...
  /dentists:
    get:
      description: Get all Dentists
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/DentistResponse'
            type: array
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get all Dentists
      tags:
      - Dentist
    post:
      description: Create a Dentist
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: DentistResponse
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/DentistPost'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/DentistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Create a Dentist
      tags:
      - Dentist
  /dentists/{id}:
    delete:
      description: Delete a Dentist
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: Dentist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Delete a Dentist
      tags:
      - Dentist
    get:
      description: Get Dentist by id
      parameters:
      - description: Dentist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DentistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Dentist by id
      tags:
      - Dentist
    patch:
      description: Patch a Dentist
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: DentistResponse
        in: body
        name: Dentist
        required: true
        schema:
          $ref: '#/definitions/DentistPatch'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DentistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Patch a Dentist
      tags:
      - Dentist
    put:
      description: Update a Dentist
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: DentistResponse
        in: body
        name: Dentist
        required: true
        schema:
          $ref: '#/definitions/DentistPut'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DentistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Update a Dentist
      tags:
      - Dentist
  /dentists/q:
    get:
      description: Get Dentist by License
      parameters:
      - description: Dentist License
        in: query
        name: license
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DentistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Dentist by License
      tags:
      - Dentist
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  APIKey:
    description: Add secret key here
    in: header
    name: SECRET_KEY
    type: apiKey
swagger: "2.0"
...